feat: Retry functionality for auth forms
