+++ src/gitClone.ts
@@ -17,6 +17,8 @@ import {
 
 import {Git} from './git';
 
+import {GitCredentialsForm} from './components/CredentialsBox'
+
 /**
  * The widget encapsulating the Git Clone UI:
  * 1. Includes the Git Clone button in the File Browser toolbar.
@@ -86,35 +88,42 @@ export class GitClone extends Widget {
     private makeApiCall(cloneUrl: string) {
         this.gitApi.clone(this.fileBrowser.model.path, cloneUrl)
             .then(async response => {
-                if (response.code != 0) {
-                    if (response.code == 128 && response.message.indexOf('could not read Username')>=0) {
-                        //request user credentials
+                let retry = false;
+                while(response.code != 0){
+                    if (response.code == 128 && (response.message.indexOf('could not read Username')>=0 || response.message.indexOf('Auth or timeout error')>=0)) {
+                        //request user credentials and try to clone again
                         const dialog = new Dialog({
                             title: 'Git credentials required',
-                            body: new GitCredentialsForm(),
-                            focusNodeSelector: 'input',
+                            body: new GitCredentialsForm('Enter credentials for remote repository', retry ? 'Incorrect username or password.' : ''),
                             buttons: [
                                 Dialog.cancelButton(),
                                 Dialog.okButton({label: 'OK'})
                             ]
                         });
+                        
+
                         const result = await dialog.launch();
                         dialog.dispose();
                         
                         if (result.button.label == 'OK') {
+                            //user accepted attempt to login
+                            //now, we can try cloning again
                             let auth = JSON.parse(decodeURIComponent(result.value));
                             //call gitApi.clone again with credentials
-                            this.gitApi.clone(this.fileBrowser.model.path, cloneUrl, auth.username, auth.password);
+                            response = await this.gitApi.clone(this.fileBrowser.model.path, cloneUrl, auth.username, auth.password);
                         }
                         else {
+                            //user cancelled attempt to log in
                             this.showErrorDialog();
+                            break;
                         }
-
+                        retry = true;
+                        
                     }
                     else {
                         this.showErrorDialog(response.message);
+                        break;
                     }
-                    
                 }
             })
             .catch(() => this.showErrorDialog())
@@ -214,54 +223,4 @@ class GitCloneForm extends Widget {
     getValue(): string {
         return encodeURIComponent(this.node.querySelector('input').value);
     }
-}
-
-/**
- * The UI for the credentials form
- */
-class GitCredentialsForm extends Widget {
-    
-    /**
-     * Create a redirect form.
-     */
-    constructor() {
-        super({node: GitCredentialsForm.createFormNode()});
-    }
-
-    private static createFormNode(): HTMLElement {
-        const node = document.createElement('div');
-        const label = document.createElement('label');
-        const user = document.createElement('input');
-        const password = document.createElement('input');
-        password.type = 'password';
-        password.id = 'git_password';
-
-        const text = document.createElement('span');
-        const warning = document.createElement('div');
-
-        node.className = 'jp-RedirectForm';
-        warning.className = 'jp-RedirectForm-warning';
-        text.textContent = 'Enter credentials for provided repository';
-        user.placeholder = 'user';
-
-        label.appendChild(text);
-        label.appendChild(user);
-        label.appendChild(password);
-        node.appendChild(label);
-        node.appendChild(warning);
-        return node;
-    }
-
-    /**
-     * Returns the input value.
-     */
-    getValue(): string {
-        let lines = this.node.querySelectorAll('input');
-        let credentials = {
-            username: lines[0].value,
-            password : lines[1].value,
-        }
-        return encodeURIComponent(JSON.stringify(credentials));
-    }
-
 }
